fix mac build warning

+++ src/Ai/Raid/SWP/Util/SWPEncounter_Kalec.h
@@ -52,8 +52,9 @@ inline constexpr uint8 KALECGOS_GROUP_COUNT = 4;
 // Rifts remain active for 10s after spawning.
 inline constexpr uint32 SPECTRAL_RIFT_ACTIVE_WINDOW_MS = 10000;
 // Used to determine if Exhaustion will expire on a bot before the rift expires. Reduced to account
-// for AI tick delay + potential latency.
-inline constexpr uint32 SPECTRAL_RIFT_ENTRY_WINDOW_MS = SPECTRAL_RIFT_ACTIVE_WINDOW_MS - 200;
+// for AI tick delay + potential latency. Signed because it is compared against an aura duration.
+inline constexpr int32 SPECTRAL_RIFT_ENTRY_WINDOW_MS =
+    static_cast<int32>(SPECTRAL_RIFT_ACTIVE_WINDOW_MS) - 200;
 inline constexpr float SPECTRAL_RIFT_SEARCH_RADIUS = 75.0f;
 inline constexpr uint32 SPECTRAL_RIFT_CACHE_INTERVAL_MS = 200;
 // Approximate Z coordinate of the Spectral Realm, used to force teleport bots down if they are
@@ -61,7 +62,7 @@ inline constexpr uint32 SPECTRAL_RIFT_CACHE_INTERVAL_MS = 200;
 inline constexpr float SPECTRAL_REALM_Z = -74.5f;
 // Curse of Boundless Agony doubles its tick damage every 5 ticks and, when removed by dispel or
 // expiration, bounces to another player. Hold off on dispelling while damage is low (until 15s).
-inline constexpr uint32 KALECGOS_DISPEL_REMAINING_MS = 15000;
+inline constexpr int32 KALECGOS_DISPEL_REMAINING_MS = 15000;
 
 inline Position const KALECGOS_TANK_POSITION =           { 1703.584f, 895.626f, 53.076f };
 inline Position const KALECGOS_INITIAL_RANGED_POSITION = { 1704.634f, 938.080f, 53.076f };
